db: Modularize database code
