+++ include/ccoin/db/blkdb.h
