+++ src/Makefile.am
@@ -9,6 +9,7 @@ brd_SOURCES=		\
 	brd.h
 
 brd_LDADD	= $(top_builddir)/lib/libccoin.la \
+		  $(top_builddir)/lib/libccoindb.la \
 		  $(top_builddir)/lib/libccoinnet.la \
 		  @GMP_LIBS@ @EVENT_LIBS@ @LMDB_LIBS@
 
@@ -18,6 +19,7 @@ picocoin_SOURCES=	\
 	wallet.c wallet.h
 
 picocoin_LDADD	= $(top_builddir)/lib/libccoin.la \
+		  $(top_builddir)/lib/libccoindb.la \
 		  $(top_builddir)/lib/libccoinnet.la \
 		  $(top_builddir)/lib/libccoinaes.la \
 		  @GMP_LIBS@ @ARGP_LIBS@ @EVENT_LIBS@ \
